+++ backend/rest_app/dependencies/auth.py
