+++ backend/rest_app/dependencies/supabase_client.py
