+++ voyago_WebApp/pages/1_Feed.py
@@ -1,7 +1,7 @@
 """Image feed view"""
 import streamlit as st
 
-from backend.rest_app.fast_api_dependencies import get_voyago
+from backend.rest_app.dependencies import get_voyago
 
 st.set_page_config(page_title="Voyago", layout="wide")
